requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7
drf-spectacular==0.26.3
python-dotenv==1.0.0